style: move tests in adequate files + skip failing test
